pyproject.toml
[tool.poetry]
name = "python-nl-to-sql"
version = "0.1.0"
description = ""
authors = ["Ricardo Cataldi <rcataldi@microsoft.com>"]
readme = "README.md"
packages = [{include = "*", from="src"}]


[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.27.0"
pydantic = "^2.8.0"
opencensus-ext-azure = "^1.1.13"
tiktoken = "^0.7.0"
asyncpg = "^0.29.0"
aistudio-request = {path = ".packages/aistudio_request-0.1.0-py3-none-any.whl"}


[tool.poetry.group.dev.dependencies]
pylint = "^2.15.9"
black = {extras = ["jupyter"], version = "^24.4.2"}
pytest = "^7.4.3"
isort = "^5.13.2"
pytest-asyncio = "^0.23.6"
python-dotenv = "^1.0.1"
ipykernel = "^6.29.3"
poetry-plugin-export = "^1.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q -s"
testpaths = ["tools",]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = ["async: marks tests as async (select with '-m async')", "slow: marks tests as slow (select with '-m slow')"]


[tool.pylint]
fail-under = 8.5
max-line-length = 100
max-complexity=10
good-names = ["i", "j", "k", "ex", "Run", "_"]
disable = ["C0114", "C0115", "C0116", "C0301", "W3101"]
load-plugins = ["pylint.extensions.mccabe", ]


[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'
exclude = '(\.git | \.hg | \.mypy_cache | \.tox | \.venv | _build | buck-out | build | dist )/'


[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 100